_requirements.txt
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
et-xmlfile==2.0.0
fastjsonschema==2.21.1
fonttools==4.57.0
fqdn==1.5.1
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
isoduration==20.11.0
jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.1
mistune==3.1.3
mizani==0.13.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
notebook-shim==0.2.4
openpyxl==3.1.5
overrides==7.7.0
pandas==2.2.3
pandocfilters==1.5.1
patsy==1.0.1
pillow==11.2.1
plotnine==0.14.5
polars==1.20.0
prometheus-client==0.21.1
python-json-logger==3.3.0
python-dateutil==2.9.0post0
pytz==2025.2
pycparser==2.22
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pyyaml==6.0.2
sniffio==1.3.1
soupsieve==2.6
statsmodels==0.14.4
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
